refactor: clean up for main layout from rebase. And changes import for react shopper hooks
