+++ src/components/cart/OrderComplete.tsx
@@ -17,7 +17,7 @@ import {
 import { OrderCompleteState } from "../checkout/types/order-pending-state";
 import { ChakraNextImage } from "../ChakraNextImage";
 import NextLink from "next/link";
-import { PresentCartState } from "@field123/epcc-react";
+import { PresentCartState } from "@elasticpath/react-shopper-hooks";
 
 interface IOrderComplete {
   state: OrderCompleteState;
