+++ src/components/checkout/types/order-pending-state.ts
@@ -1,4 +1,4 @@
-import { PresentCartState } from "@field123/epcc-react";
+import { PresentCartState } from "@elasticpath/react-shopper-hooks";
 import { CheckoutForm as CheckoutFormType } from "../form-schema/checkout-form-schema";
 import { ConfirmPaymentResponse } from "@moltin/sdk";
