+++ src/components/layouts/MainLayout.tsx
@@ -4,15 +4,23 @@ import Footer from "../footer/Footer";
 import type { ReactNode } from "react";
 import type { NavigationNode } from "../../lib/build-site-navigation";
 import { Toaster } from "../toast/toaster";
+import Head from "next/head";
 
 interface IMainLayout {
-  nav?: NavigationNode[];
+  nav: NavigationNode[];
   children: ReactNode;
 }
 
-const MainLayout = ({ nav = [], children }: IMainLayout): JSX.Element => {
+const MainLayout = ({ nav, children }: IMainLayout): JSX.Element => {
   return (
     <>
+      <Head>
+        <title>D2C Starter Kit</title>
+        <meta
+          name="description"
+          content="D2C Starter Kit - a store front starter for Elastic Path Commerce Cloud"
+        />
+      </Head>
       <Toaster />
       <Header nav={nav} />
       {children}
