+++ src/lib/resolve-shopping-cart-props.ts
@@ -1,6 +1,6 @@
-import { CartState, PresentCartState } from "@field123/epcc-react";
+import { CartState, PresentCartState } from "@elasticpath/react-shopper-hooks";
 import { ICart } from "../components/cart/Cart";
-import { getPresentCartState } from "@field123/epcc-react";
+import { getPresentCartState } from "@elasticpath/react-shopper-hooks";
 
 export function resolveShoppingCartProps(
   state: CartState,
