+++ src/lib/store-wrapper-ssr.ts
@@ -2,7 +2,7 @@ import { GetServerSidePropsContext, GetServerSidePropsResult } from "next";
 import { ParsedUrlQuery } from "querystring";
 import { buildSiteNavigation, NavigationNode } from "./build-site-navigation";
 import { getCart } from "../services/cart";
-import { StoreContextSSR } from "@field123/epcc-react";
+import { StoreContextSSR } from "@elasticpath/react-shopper-hooks";
 import { getCartCookie } from "./cart-cookie";
 import { getEpccImplicitClient } from "./epcc-implicit-client";
