+++ src/pages/_app.tsx
@@ -4,7 +4,7 @@ import { ChakraProvider } from "@chakra-ui/react";
 import theme from "../styles/theme";
 import "focus-visible/dist/focus-visible";
 import "../styles/globals.css";
-import { StoreContext } from "@field123/epcc-react";
+import { StoreContext } from "@elasticpath/react-shopper-hooks";
 import { NextPage } from "next";
 import { ReactElement, ReactNode } from "react";
 import { getMainLayout } from "../lib/get-main-layout";
