+++ src/pages/cart.tsx
@@ -1,7 +1,7 @@
 import { Box, Heading, Image } from "@chakra-ui/react";
 import type { NextPage } from "next";
 import { withStoreServerSideProps } from "../lib/store-wrapper-ssr";
-import { useCart } from "@field123/epcc-react";
+import { useCart } from "@elasticpath/react-shopper-hooks";
 import Cart from "../components/cart/Cart";
 import { resolveShoppingCartProps } from "../lib/resolve-shopping-cart-props";
 import { globalBaseWidth } from "../styles/theme";
